created delete.mjs
function that retrieves an link/cta for a post, that let's you delete the post. -> it will loop through all the delete links, and then adds a eventListener on "click" where it retrieves the ID from the clicked element that should be the same ID as a created post. It then uses the ID from the clicked element and passes it into a API endpoint, which then is used as a argument for a delete fetch call